[virtual] Unlock static build support
