+++ src/gui/CMakeLists.txt
@@ -1550,6 +1550,7 @@ if (FORCE_STATIC_LIBS)
     target_link_libraries(qgis_gui
       provider_spatialite_gui_a
       provider_wfs_gui_a
+      provider_virtuallayer_gui_a
     )
   endif()
   if (HAVE_POSTGRESQL)
