+++ src/providers/virtual/qgsvirtuallayerprovider.h
@@ -1,5 +1,5 @@
 /***************************************************************************
-           qgsvirtuallayerprovider.cpp Virtual layer data provider
+           qgsvirtuallayerprovider.h Virtual layer data provider
 begin                : Jan, 2015
 copyright            : (C) 2015 Hugo Mercier, Oslandia
 email                : hugo dot mercier at oslandia dot com
@@ -25,9 +25,6 @@ email                : hugo dot mercier at oslandia dot com
 #include "qgsvirtuallayersqlitehelper.h"
 
 #include "qgsprovidermetadata.h"
-#ifdef HAVE_GUI
-#include "qgsproviderguimetadata.h"
-#endif
 
 class QgsVirtualLayerFeatureIterator;
 
@@ -61,6 +58,8 @@ class QgsVirtualLayerProvider final: public QgsVectorDataProvider
     QgsAttributeList pkAttributeIndexes() const override;
     QSet<QgsMapLayerDependency> dependencies() const override;
     bool cancelReload() override;
+
+    static QString providerKey();
 
   private:
 
@@ -136,15 +135,6 @@ class QgsVirtualLayerProviderMetadata final: public QgsProviderMetadata
     QgsVirtualLayerProviderMetadata();
     QgsVirtualLayerProvider *createProvider( const QString &uri, const QgsDataProvider::ProviderOptions &options, QgsDataProvider::ReadFlags flags = QgsDataProvider::ReadFlags() ) override;
 };
-
-#ifdef HAVE_GUI
-class QgsVirtualLayerProviderGuiMetadata final: public QgsProviderGuiMetadata
-{
-  public:
-    QgsVirtualLayerProviderGuiMetadata();
-    QList<QgsSourceSelectProvider *> sourceSelectProviders() override;
-};
-#endif
 
 // clazy:excludeall=qstring-allocations
